+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Features
 * User can delete his account from his profile page
 * Add new component to enforce password strength
-* Add explicit error messages on 
+* Add constraint on password confirmation
 * Add rating form
 ### Fix
 ## Version 1.4.3